Add LatticeManager — main boilerplate for Harrington impl
Basically all the Harrington rules will be implemented through
LatticeManager (at least the way I see it right now). Broadcasting and
all other roles can probably be handled through it. Maybe this should be
refactored at some point, but I think this is the most straightforward
impl to start with.