add basic implementations for LatticeManager and Lattice
I created three levels of lattices in order to get around the problem with having every lattice field be an Option type. All of the automata and edges are now owned by LatticeManager, in addition to the lattices. This is to resolve the issue of what owns the edges between BaseLattices. The primary and dual fields in LatticeManager own the Edges and map the Automata to their edges. The BaseLattice just has a HashMap from points to AutomatonId's, which are used to access the fields in LatticeManager. I also added some basic structure for the HarringtonRules to build on. Some of the creation/building of the LatticeManager and all of its children still needs to be implemented